src/frontend/requirements.txt
aio-pika==9.4.3
streamlit==1.39.0
pyngrok==7.2.0
SQLAlchemy==2.0.30
python-dotenv==1.0.1
plotly==5.24.1
asyncpg==0.29.0
streamlit-authenticator==0.4.1
opencv-python==4.10.0.84
